Refactored a command into separete file
