+++ src/civilization/console/mod.rs
@@ -10,3 +10,4 @@ mod start_command;
 mod player_end_building_command;
 mod eliminate_city_command;
 mod stupid_ai_command;
+mod make_a_move_command;
